changed dir/file structure, added fonts to the project
