--- src/main/java/com/tis/dbf/model/AcademicYear.java
@@ -1,20 +1,0 @@
-package com.tis.dbf.model;
-
-import jakarta.xml.bind.annotation.*;
-import lombok.Data;
-
-import java.util.List;
-
-@Data
-@XmlAccessorType(XmlAccessType.FIELD)
-public class AcademicYear {
-
-    @XmlElement(name = "Registration.Date")
-    private String registrationDate;
-
-    @XmlElement(name = "Years")
-    private String years;
-
-    @XmlElement(name = "Subjects")
-    private StudySubjects studySubjects;
-}
