--- src/main/resources/StudiesTEST.xml
@@ -1,100 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-<Studies>
-	<Study>
-		<id.UPN>42853223</id.UPN>
-		<Study.Programme>Aplikovaná ekonómia</Study.Programme>
-		<Study.Field>Ekonómia a manažment</Study.Field>
-		<Faculty>Fakulta sociálnych a ekonomických vied</Faculty>
-		<Level>Prvý</Level>
-		<Degree>Bc</Degree>
-		<Standard.Length>3</Standard.Length>
-		<Study.Model>BC3</Study.Model>
-		<Language>Slovenský</Language>
-		<Form>Denná</Form>
-		<Study.Admission>
-			<Type>Prijímacie konanie</Type>
-			<Date>5.06.2001</Date>
-		</Study.Admission>
-		<Total.Credits>180</Total.Credits>
-		<Academic.Years>
-			<AcademicYear>
-				<Registration.Date>12.09.2001</Registration.Date>
-				<Years>2001-2002</Years>
-				<Subjects>
-					<Subject>
-						<UIDP>SAAV</UIDP>
-						<Grade>B</Grade>
-						<Type>P</Type>
-						<Attempt>Riadny</Attempt>
-						<End.Subject>Skúška</End.Subject>
-						<End.Subject.Date>11.01.2001</End.Subject.Date>
-						<Credits>6</Credits>
-						<Semester>Z</Semester>
-					</Subject>
-				</Subjects>
-			</AcademicYear>
-			<AcademicYear>
-				<Registration.Date>12.09.2002</Registration.Date>
-				<Years>2002-2003</Years>
-				<Subjects>
-				</Subjects>
-			</AcademicYear>
-			<AcademicYear>
-				<Registration.Date>15.09.2003</Registration.Date>
-				<Years>2003-2004</Years>
-				<Subjects>
-				</Subjects>
-			</AcademicYear>
-			<AcademicYear>
-				<Registration.Date>10.09.2004</Registration.Date>
-				<Years>2004-2005</Years>
-				<Subjects>
-				</Subjects>
-			</AcademicYear>
-		</Academic.Years>
-		<Interruptions>
-			<Interruption>
-				<Reason>Rodinné dôvody</Reason>
-				<Start.Date>01.03.2003</Start.Date>
-				<End.Date>31.08.2003</End.Date>
-			</Interruption>
-			<Interruption>
-				<Reason>Rodinné dôvody</Reason>
-				<Start.Date>01.03.2004</Start.Date>
-				<End.Date>31.08.2005</End.Date>
-			</Interruption>
-			<Interruption>
-				<Reason>Zdravotne dôvody</Reason>
-				<Start.Date>01.03.2005</Start.Date>
-				<End.Date>31.08.2006</End.Date>
-			</Interruption>
-		</Interruptions>
-		<Abroad.Programmes>
-			<Abroad.Programme>
-				<Start.Date>15.03.2004</Start.Date>
-				<End.Date>15.06.2004</End.Date>
-				<Semester>2L</Semester>
-				<University>University of Barcelona</University>
-				<Country>Španielsko</Country>
-			</Abroad.Programme>
-		</Abroad.Programmes>
-		<Study.End>
-			<State.Exams>
-				<State.Exam>
-					<UIDP>SAAAD</UIDP>
-					<Date>17.06.2005</Date>
-					<Grade>A</Grade>
-					<Credits>8</Credits>
-				</State.Exam>
-			</State.Exams>
-			<Thesis>
-				<NameSK>Nezamestnanosť v regiónoch Slovenska</NameSK>
-				<NameENG>Unemployment in the Regions of Slovakia</NameENG>
-				<Grade>B</Grade>
-				<Defence.Date>02.06.2005</Defence.Date>
-			</Thesis>
-			<Rector.FullName>Janko Mrkvička</Rector.FullName>
-		</Study.End>
-		<Study.Status>Absolvované</Study.Status>
-	</Study>
-</Studies>
