--- src/main/resources/example.xml
@@ -1,14 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-<Predmety>
-    <Predmet>
-        <UIDP>SAAA</UIDP>
-        <ID.system1/>
-        <ID.system2/>
-        <ID.system3/>
-        <Name.SK>Povinné predmety</Name.SK>
-        <Name.EN/>
-        <Name.Part/>
-        <exc/>
-        <acc/>
-    </Predmet>
-</Predmety>
