+++ src/main/resources/fxml/DownloadScene.fxml
@@ -11,10 +11,8 @@
 <?import java.net.URL?>
 <?import javafx.scene.layout.StackPane?>
 <?import javafx.scene.control.PasswordField?>
-<AnchorPane fx:id="anchorPane" prefWidth="1920" prefHeight="1080" xmlns:fx="http://javafx.com/fxml" fx:controller="com.controllers.controllers.DownloadController">
-    <stylesheets>
-        <URL value="@styles.css" />
-    </stylesheets>
+<?import java.lang.String?>
+<AnchorPane stylesheets="@styles.css" fx:id="anchorPane" xmlns:fx="http://javafx.com/fxml" fx:controller="controllers.DownloadController">
 
     <StackPane fx:id="stackPane" AnchorPane.leftAnchor="0" AnchorPane.rightAnchor="0" AnchorPane.topAnchor="0" AnchorPane.bottomAnchor="0">
         <Label fx:id="statusLabel" text="Initializing..." styleClass="status-label"/>
