+++ src/main/resources/fxml/Login.fxml
@@ -8,13 +8,9 @@
 <?import javafx.scene.image.ImageView?>
 <?import javafx.scene.layout.GridPane?>
 <?import javafx.scene.layout.StackPane?>
-
-<?import java.net.URL?>
-<StackPane xmlns="http://javafx.com/javafx" xmlns:fx="http://javafx.com/fxml"
-           fx:controller="com.controllers.controllers.LoginController">
-    <stylesheets>
-        <URL value="@styles.css"/>
-    </stylesheets>
+<?import java.lang.String?>
+<StackPane stylesheets="@styles.css" xmlns="http://javafx.com/javafx" xmlns:fx="http://javafx.com/fxml"
+           fx:controller="controllers.LoginController">
 
     <GridPane alignment="CENTER" hgap="10" vgap="10">
         <!-- Logo - does not block other rows -->
@@ -22,9 +18,7 @@
             <GridPane.margin>
                 <Insets bottom="-30"/> <!-- Negative bottom margin to "pull" fields closer -->
             </GridPane.margin>
-            <image>
-                <Image url="@/uk_logo.png"/>
-            </image>
+            <Image url="@/uk_logo.png"/>
         </ImageView>
 
         <!-- Username field -->
